# api_gateway/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GDI MCP Server API",
    "version": "1.0.0",
    "description": "API de Gestión Documental Inteligente para gobiernos latinoamericanos. Permite consultar expedientes, documentos, firmas pendientes y catálogos del sistema."
  },
  "servers": [
    {
      "url": "__OPENAPI_BASE_URL__"
    }
  ],
  "paths": {
    "/api/v1/cases/search": {
      "get": {
        "operationId": "searchCases",
        "summary": "Buscar expedientes",
        "description": "Busca expedientes por número, referencia, contenido de documentos, nombres, direcciones, etc.",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Texto a buscar (número, referencia, contenido)"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "inactive",
                "archived"
              ]
            },
            "description": "Estado del expediente"
          },
          {
            "name": "date_filter",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "hoy",
                "ayer",
                "ultimos_7_dias",
                "ultimos_30_dias"
              ]
            },
            "description": "Filtro temporal"
          },
          {
            "name": "sector_filter",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filtrar por sector (acronym)"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Página"
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Resultados por página (max 100)"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de expedientes con case_number, reference, ai_summary"
          },
          "401": {
            "description": "No autorizado"
          }
        },
        "security": [
          {
            "OAuth2": [
              "openid",
              "profile",
              "email"
            ]
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/api/v1/cases/{case_id}": {
      "get": {
        "operationId": "getCase",
        "summary": "Detalle de expediente",
        "description": "Obtiene información completa de un expediente específico",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "UUID del expediente"
          },
          {
            "name": "include_documents",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Incluir lista de documentos"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle del expediente"
          },
          "404": {
            "description": "Expediente no encontrado"
          }
        },
        "security": [
          {
            "OAuth2": [
              "openid",
              "profile",
              "email"
            ]
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
    "/api/v1/documents/pending-signatures": {
      "get": {
        "operationId": "getPendingSignatures",
        "summary": "Firmas pendientes",
        "description": "Lista documentos que esperan la firma del usuario actual",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de documentos pendientes de firma"
          },
          "401": {
            "description": "No autorizado"
          }
        },
        "security": [
          {
            "OAuth2": [
              "openid",
              "profile",
              "email"
            ]
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
    "/api/v1/cases/": {
      "post": {
        "operationId": "createCase",
        "summary": "Crear expediente",
        "description": "Crea un nuevo expediente con caratula automatica",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "case_template_id",
                  "reference"
                ],
                "properties": {
                  "case_template_id": {
                    "type": "string",
                    "description": "UUID del template"
                  },
                  "reference": {
                    "type": "string",
                    "description": "Asunto del expediente"
                  },
                  "owner_sector_id": {
                    "type": "string",
                    "description": "UUID del sector propietario (opcional)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Expediente creado"
          },
          "400": {
            "description": "Datos invalidos"
          }
        },
        "security": [
          {
            "OAuth2": [
              "openid",
              "profile",
              "email"
            ]
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
    "/api/v1/system/users/list": {
      "get": {
        "operationId": "listAllUsers",
        "summary": "Listar todos los usuarios",
        "description": "Obtiene lista paginada de todos los usuarios del tenant",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Pagina"
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            },
            "description": "Resultados por pagina"
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Buscar por nombre o email"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuarios"
          }
        },
        "security": [
          {
            "OAuth2": [
              "openid",
              "profile",
              "email"
            ]
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
    "/api/v1/records/{record_id}/cases": {
      "get": {
        "operationId": "getRecordCases",
        "summary": "Expedientes vinculados al legajo",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de expedientes vinculados"
          }
        },
        "security": [
          {
            "OAuth2": [
              "openid",
              "profile",
              "email"
            ]
          },
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "post": {
        "operationId": "linkRecordCase",
        "summary": "Vincular expediente al legajo",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Expediente vinculado"
          }
        },
        "security": [
          {
            "OAuth2": [
              "openid",
              "profile",
              "email"
            ]
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/api/v1/records/{record_id}/cases/{link_id}": {
      "delete": {
        "operationId": "unlinkRecordCase",
        "summary": "Desvincular expediente del legajo",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "link_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Expediente desvinculado"
          }
        },
        "security": [
          {
            "OAuth2": [
              "openid",
              "profile",
              "email"
            ]
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "OAuth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://__OPENAPI_AUTH0_DOMAIN__/authorize",
            "tokenUrl": "https://__OPENAPI_AUTH0_DOMAIN__/oauth/token",
            "scopes": {
              "openid": "OpenID Connect",
              "profile": "User profile",
              "email": "User email"
            }
          }
        }
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key para autenticación (fallback)"
      }
    }
  }
}